Custom logger responsability implementation moved to the client. Logger more flexible.
