--- logger/include/LoggerToFile.h
@@ -1,6 +1,0 @@
-#ifndef LOGGER_TO_FILE_H
-#define LOGGER_TO_FILE_H
-
-// TODO
-
-#endif
