--- logger/src/LoggerToFile.cpp
@@ -1,3 +1,0 @@
-#include "LoggerToFile.h"
-
-// TODO
