+++ logger/src/ScribeAPI.cpp
@@ -3,13 +3,10 @@
 #include <string>
 
 #include "ScribeAPI.h"
-#include "ILogger.h"
-#include "LoggerToStdOut.h"
 
 namespace scribe
 {
-    // TODO: Refactor to initialize logger.
-    std::shared_ptr<ILogger> logger_ = std::make_shared<LoggerToStdOut>();
+    std::shared_ptr<ILogger> logger_;
 
     void logVerbose(const char* file, const char* function, const int line, const std::string& message)
     {
@@ -34,5 +31,22 @@ namespace scribe
     void logCritical(const char* file, const char* function, const int line, const std::string& message)
     {
         logger_.get()->logCritical(file, function, line, message);
-    }            
+    }
+    
+    void initializeLogger(std::shared_ptr<ILogger> logger)
+    {
+        // TODO dynamic cast.
+        logger_ = std::move(logger);
+    }
+
+    void configMinimunLoggerLevel(std::shared_ptr<ILogger> logger, scribe::LOG_LEVEL loggerLevel)
+    {
+        logger->setMinimunLoggerLevel(loggerLevel);
+    }
+
+    // void finalizeLogger(std::shared_ptr<ILogger>& logger)
+    // {
+    //      // TODO
+    // }    
 }
+
